Update REAME, split tests.
